Add cross-platform support

+++ src/audio/heapinit.c
@@ -23,7 +23,7 @@
 
 void alHeapInit(ALHeap *hp, u8 *base, s32 len)
 {
-    s32 extraAlign = (AL_CACHE_ALIGN+1) - ((s32) base & AL_CACHE_ALIGN);
+    s32 extraAlign = (AL_CACHE_ALIGN+1) - ((intptr_t) base & AL_CACHE_ALIGN);
     
     if (extraAlign != AL_CACHE_ALIGN+1)
         hp->base = base + extraAlign;
